pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orbbec-astra-raw"
version = "0.1.0"
description = "Unofficial pure-Python driver for the Orbbec Astra Pro IR/depth sensor"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["orbbec", "astra", "depth", "ir", "usb", "driver", "lidar"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Video :: Capture",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "pyusb>=1.2",
    "numpy>=1.21",
]

[project.optional-dependencies]
viewer = ["opencv-python>=4.5"]
image  = ["Pillow>=9.0"]
all    = ["opencv-python>=4.5", "Pillow>=9.0"]

[project.scripts]
astra-ir-view = "astra_raw.cli:cmd_view"
astra-ir-save = "astra_raw.cli:cmd_save"
astra-ir-dump = "astra_raw.cli:cmd_dump"

[project.urls]
Homepage   = "https://github.com/chintan-27/orbbec-astra-raw"
Issues     = "https://github.com/chintan-27/orbbec-astra-raw/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/astra_raw"]
